requirements/requirements-new.txt
pillow
tqdm
matplotlib
pandas
numpy
seaborn
plotly
statsmodels
pdfplumber
scikit-learn
opencv-python
scikit-image
librosa
soundfile
sounddevice
albumentations
h3
shapely
segmentation-models-pytorch
ultralytics
emoji
regex
nltk
pymorphy3
textdistance
sentence-transformers
augmentex
rank-bm25
ollama
noisereduce
catboost
lightgbm
xgboost
optuna
shap
transformers
datasets
python-multipart
fastapi
uvicorn
requests
pytesseract
doclayout-yolo
pdf2image
faster-whisper
silero-vad
jiwer
streamlit
geopandas
folium
streamlit-folium
joblib
gpxpy
osmnx
contextily
rasterio
h5py
aiogram
pymupdf
langchain
langchain-community
faiss-cpu
python-docx
imbalanced-learn


---
Не забыть поставить торч под куду
pip uninstall torch torchvision
pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu128